requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
webuiapi==0.9.5
requests
loguru
retry
pyyaml